requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
google-generativeai==0.3.2
cohere==4.37
transformers==4.36.0
torch==2.1.0
networkx==3.2.1
numpy==1.26.2
